unit tests from optimizer.rs moved to tests/unit/optimizer.rs
